Approx entropy and pioncare changes
updated the swan template to include approx entropy for both the RR
intervals and the Ttot. Updated the poincares so it will give both RR
and Ttot plots. SD1 and SD2 are now displayed right on the graph, and
again below. the results are in seconds and are rounded to 3 places
after the decimal.