Edge event detection and interval correction
noticed that intervals were being calculated in the wrong place if the
peaks or bursts are being filtered. Moved around burst detect things so
it does it correctly.
also added edge event flag as bol term. True == Edge event.
made a quick and dirty area under the curve for callable ROI.
also, fired peak to peak intervals to be in time not index.
#26372719814005