error fixed
settings save and load added.
got around the error by using try/except to pass through the function.
it appears that the problem is because i am passing an empty data frame
using .tolist() into the ttotal and edge case functions, which can’t
handle that. it is now sanitized for my protection. I may need to do
this for other functions too.

#27065940339996
#16683064948018