rolling baseline
fixed a typo with the Baseline-Rolling key

also rearranged the burst detection so that the filter doesn’t create
problems with burst amplitudes.

lastly, added an early return to the burst detection so that if there
are no bursts after the filter, it returns the empty data frame and the
failure before trying to do any of the other measurements.

#26372719814005