peak maximum fix, attack, decay
fixed peak maximum. separated it from the peaks per burst loop. also
moved the try-except to inside the for loop, so that one failure does
not cause all of the burst to fail.
also fixed the ‘nan’ error. it is now, correctly, nan
lastly, added attack and decay to the burst results tables.
#27489202151582
#27489202151588