lin subtraction SW correction
noticed that the lin subtraction was not behaving correctly for seans
data because it is normalized. added a parameter about this called
relative baseline. for non-norm data = 0, norm = 1.
added filter for minimum number of peaks per burst. for swans data,
will be 1.
i also suggest that lin sub r = 0.85 for sean’s data.
#27611837042208
#27611837042205