lin subtraction with smoothing
tried to add rolling mean to the linear subtraction function, to see if
it would improve the r value of the fit line. in some cases, like the
edge case we care about, it does. so, I will add a filter that on
chooses which method is the better one for lin fit. still must be
greater than some thresh for perform thing anyways.
setting smoothing window to 50, arbitrarily, for sean.
#27611837042205