refinement
removed division import, i’m not using it.
setting for milliseconds is now stored in settings, but is only applied
in the ‘Morgan’ File type.
changed rolling baseline window from milliseconds to seconds for
consistency.
added print statements for when a time series has no peaks.
removed Results[‘Peak Summary’], which is now better represented by
Results[‘Peaks-Master Summary’]
added print statements for when a time series has no bursts.
updated histent_wrapper to include Data in the inputs, since it’s
column names are used.
added the ‘all’ function to histent, which allows for the function to
loop through all available measurements for an event type.
moving_statistics added: Generates the moving mean, standard deviation,
and count for a given measurement.
Saves out the dataframes of these three results automatically with
the window size in the name.
If meas == 'All', then the function will loop and produce these
tables for all measurements.
added analyze, which is the function that defines the pipeline for
event detection. Follows the strict '3 arguments in, 3 arguments out'
rule. These dictionaries must be intalized before this, as well as
all of the Settings values.
Detects bursts and peaks for the data file that is uploaded.
Settings used are saved out automatically with a time stamp as a
receipt of each analysis performed. It has a few handy printed outputs,
such as how long an analysis took, which step was just completed, lists
of which objects contained no peaks or bursts. it also prints a list of
key names and analysis measurements, which can be used in further
analysis steps.
#28634956844599
#26645061569222
#27890282380033